requirements.txt
pandas==1.4.3
pyodbc==4.0.34
xlrd==2.0.1
pymssql==2.2.5
openpyxl==3.0.10
xlsxwriter==3.0.3
numpy==1.23.1
sqlalchemy==1.4.39
python-dotenv==0.20.0
selenium
datetime
webdriver_manager